Add tests (happy & sad) for DELETE request
